lxd-to-incus: Split snap source to separate file
